==> code/1-generar_dataset.ipynb <==
# Process Monitoring dataset

import pandas as pd

# cargar datos en bruto
monitoringdf = pd.read_csv('../data/_Process_monitoring__202303251137.csv', sep=',')

# Creacion del las columnas datetime y latitude_longitude
monitoringdf['datetime'] = monitoringdf['date'] + ' ' + monitoringdf['time']
monitoringdf['latitude_longitude'] = monitoringdf.apply(lambda row: (row['latitude'], row['longitude']), axis=1)

# eliminar las columnas latitude, longitude, time, velocity. Luego reordenar columnas
monitoringdf.drop(columns=['latitude', 'longitude','time'], inplace=True)
monitoringdf = monitoringdf[['id', 'date','datetime', 'latitude_longitude', 'lap', 'navigation', 'vehicle_id']]
monitoringdf

### Cleaning

# DUPLICADOS
print(f'Shape dataset: {monitoringdf.shape}')

# Verificar y contar si hay puntos duplicados segun datetime y latitude_longitude
base = ['datetime', 'latitude_longitude']
twocols_dupli = monitoringdf.duplicated(subset=base)
print(f"Nro de duplicados en {base}: {twocols_dupli.sum()}")

# ============================ Solo para comprobacion ============================
# Verificar y contar si hay puntos duplicados segun latitude_longitude
# onecol_dupli =  monitoringdf['latitude_longitude'].duplicated()
# print(len(twocols_dupli), len(onecol_dupli))
# print(twocols_dupli.sum(), onecol_dupli.sum())
# print(abs(twocols_dupli.sum() - onecol_dupli.sum()))
# tmp = [ not x==y for x,y in zip(twocols_dupli, onecol_dupli)]
# data_tmp = monitoringdf.copy()
# data_tmp['dupli'] = tmp
# data_tmp[data_tmp['dupli'] == True]
# ================================================================================

# Eliminar columnas
monitoringdf.drop_duplicates(subset=base, inplace=True)
print(f"Duplicados eliminados => shape dataset: {monitoringdf.shape}")

# NANs
# verificar si hay datos faltantes
print("Cantidad de NaNs:")
for col in monitoringdf:
    nans = monitoringdf[col].isna().sum()
    print(f"\tColumna {col}: {nans}")

# Bus Stops Ida dataset

# leer dataset busstop
busstopdf = pd.read_csv('../data/busstops_ida.csv')

# diseño del df: creacion de latitude_longitude
busstopdf['latitude_longitude'] = busstopdf.apply(lambda row: (row['latitude'], row['longitude']), axis=1)
busstopdf.drop(columns=['number', 'latitude', 'longitude'], inplace=True)

# para fines practicos modificamos el id = index para mejor lectura
busstopdf['id'] = [f'{i+100}' for i in busstopdf.index]

# mostrar info y elementos
print(busstopdf.info())
busstopdf

### Crear links

# creamos dataframe links y renombramos columnas
tmp1 = busstopdf[['id','name']].head(busstopdf.shape[0]-1).reset_index(drop=True)
# tmp1.rename(columns={'id':'ini_id', 'name':'ini_name'}, inplace=True)
tmp1.rename(columns={'id':'init_stop', 'name':'init_name'}, inplace=True)

tmp2 = busstopdf[['id','name']].tail(busstopdf.shape[0]-1).reset_index(drop=True)
# tmp2.rename(columns={'id':'end_id', 'name':'end_name'}, inplace=True)
tmp2.rename(columns={'id':'end_stop', 'name':'end_name'}, inplace=True)
linksdf = pd.concat([tmp1, tmp2], axis=1)

# creamos la columnas link
# linksdf['link'] = linksdf.apply(lambda row: f"{row['ini_id']}:{row['end_id']}", axis=1)
linksdf['link'] = linksdf.apply(lambda row: f"{row['init_stop']}:{row['end_stop']}", axis=1)
linksdf['id_link'] = linksdf.index
linksdf.insert(0, 'id_link', linksdf.pop('id_link'))
linksdf

# Calcular Travel Time

from haversine import haversine as hvs, Unit
import numpy as np

# FUNCION: filtrar los puntos de ida
def idaPoints(df):
    end1_stop = {'latitude_longitude':(-13.5513636,-71.988304), 'navigation':75, 'radio':50}
    end2_stop = {'latitude_longitude':(-13.5511395,-71.986925), 'navigation':206, 'radio':50}
    for i, row in df.iterrows():
        if (itPassed(row, end1_stop) or itPassed(row, end2_stop)):
            break
    return df.head(i+1)

# FUNCION: ver si un punto paso por un stop
def itPassed(point, stop):
    nav = abs(point['navigation'] - stop['navigation'])
    dis = round(hvs(point['latitude_longitude'], stop['latitude_longitude'], unit=Unit.METERS), 2)
    return (nav<=45 and dis<=stop['radio'])

# FUNCION: 
def searchStopsPoints(row, stops, r_nav=45, r_dtec=60):
    # verificar la navegacion y el radio de distancia
    stops = stops[abs(stops['navigation'] - row['navigation']) <= r_nav]
    stops['dis'] = stops['latitude_longitude'].apply(lambda stop: round(hvs(row['latitude_longitude'], stop, unit=Unit.METERS), 2))
    stops = stops[stops['dis'] <= r_dtec]
    
    # preparar el output
    row['stop'] = str(stops.loc[stops['dis'].idxmin(), 'id']) if len(stops) != 0 else np.nan
    row['dis'] = float(stops.loc[stops['dis'].idxmin(), 'dis']) if len(stops) != 0 else 0    
    return row

# FUNCION: CALCULA EL TIEMPO EN SEG DE UN PUNTO CON SU ANTERIOR
def time_travel(row, df):
    if {row['init_stop'], row['end_stop']}.issubset(set(df.index)):
        # print('if')
        ti, te =  df.loc[row['init_stop'], 'datetime'], df.loc[row['end_stop'], 'datetime']
        row['date_time_init'] = ti
        row['date_time_end'] = te
        time = abs((ti - te).total_seconds())
        
        #row['time_travel'] =  time if time < 300 else np.nan
        row['time_travel'] =  time
        row['init_idp'], row['end_idp'] = df.loc[row['init_stop'], 'id'], df.loc[row['end_stop'], 'id']
        row['veh'] = df.loc[row['init_stop'], 'vehicle_id']
        row['lap'] = df.loc[row['init_stop'], 'lap']
        row['date'] = df.loc[row['init_stop'], 'date']

    else:
        # print('else')
        row['date_time_init'] = np.nan
        row['date_time_end'] = np.nan
        row['time_travel'] = np.nan
        row['init_idp'] = np.nan
        row['end_idp'] = np.nan
        row['veh'] = np.nan
        row['lap'] = np.nan
        row['date'] = np.nan
        
    return row

### Crear samples y calcular travels time

# ordenar df por fechas
monitoringdf['datetime'] = pd.to_datetime(monitoringdf['datetime'], format='%Y-%m-%d %H:%M:%S')
monitoringdf = monitoringdf.sort_values(by='datetime')
print("Type columna 'datetime' : ", monitoringdf['datetime'].dtype)
monitoringdf.head()

# preparar el dataset para iterar en todos los posibles samples
monitoringdf_tmp = monitoringdf.copy()
#monitoringdf_tmp = monitoringdf_tmp[(monitoringdf_tmp['date'] == '2023-03-01')] #| (monitoringdf_tmp['date'] == '2023-03-02')]
monitoringdf_tmp['sample'] = monitoringdf.apply(lambda row: f"{row['date']} {row['vehicle_id']} {row['lap']}" , axis=1)
n_posibles_samples = len(monitoringdf_tmp['sample'].unique())
print('nro posibles samples:', n_posibles_samples)

# iterar en todos los posibles
n_descartados = 0
id_sample = 0
list_err = []
list_samplesdf = []
date = ''
for ids, pointsdf in monitoringdf_tmp.groupby('sample'):
    # Seguir procesamiento
    if date != pointsdf['date'].unique()[0]:
        date = pointsdf['date'].unique()[0]
        print(date)

    # ordenar el sample por datetime
    pointsdf.sort_values(by='datetime', inplace=True)

    # identificar los puntos de ida del sample
    idaPointsdf = idaPoints(pointsdf.reset_index(drop=True))

    # verificar puntos por stop
    stopsPointsdf = idaPointsdf.apply(lambda row: searchStopsPoints(row, busstopdf), axis=1)
    stopsPointsdf = stopsPointsdf[stopsPointsdf['stop'].notna()].reset_index()

    # Seleccionar un unico punto por stops
    stopsPointsdf.sort_values(by=['stop', 'dis'], ascending=True, inplace=True, ignore_index=True, key=np.int64)
    stopsPointsdf.drop_duplicates(subset=['stop'], keep='first', inplace=True, ignore_index=True)

    # filtrar los stop sin puntos 
    missing_stops = list(set(busstopdf['id']) - set(stopsPointsdf['stop'].unique()))

    # verificar la cantidad de puntos filtrados
    if len(missing_stops) < 24:
        # verificar la secuencialidad de los datetime
        byS = list(stopsPointsdf.sort_values(by='stop')['stop'])
        byT = list(stopsPointsdf.sort_values(by='datetime')['stop'])
        if byS == byT:
            # crear sample y calcular el travel_time
            sampledf = linksdf.copy()
            stopsPointsdf.set_index('stop', inplace=True)
            sampledf = sampledf.apply(lambda row: time_travel(row, stopsPointsdf), axis=1)
            
            # crear el id_sample
            # sampledf['id_sample'] = 
            sampledf.insert(0, 'id_sample', [id_sample] * sampledf.shape[0])
            list_samplesdf.append(sampledf)
            id_sample += 1
        else:
            # registrar sample defectoso
            list_err.append(stopsPointsdf)
            print(' ', ids)
    else:
        n_descartados += 1

# mostrar nro de samples encontrados
print('='*100)
print(f"{len(list_samplesdf)} Samples buenos")
print(f"{len(list_err)} Samples con errores de continuidad de tiempo")
print(f"{n_descartados} Samples descartados por falta de puntos")

# Reparar samples con error de secuencialidad de tiempo
# (los samples reparados se debe agregar a 'list_samplesdf')

samples_nans = pd.concat(list_samplesdf)
samples_nans

# Convertir a csv
pd.concat(list_samplesdf).to_csv('../data/samples_nans.csv', index=False)
