==> code/2-preprocesar_dataset.ipynb <==
import pandas as pd

# leer dataset 
samples_nans = pd.read_csv('../data/samples_nans.csv', sep=',')
samples_nans['date_time_init'] = pd.to_datetime(samples_nans['date_time_init'])
samples_nans['date_time_end'] = pd.to_datetime(samples_nans['date_time_end'])
samples_nans

import numpy as np

can_nans = samples_nans['time_travel'].isna().sum()
print(f"{can_nans} nans de {samples_nans.shape[0]} rows equivale {can_nans*100/samples_nans.shape[0]:0.0f}%")

### Reemplazar NaNs con promedios

from more_itertools import consecutive_groups

def intervals(numeros):
    intervals = []
    for group in consecutive_groups(numeros):
        group = list(group)
        if len(group) == 1:
            intervals.append(tuple(group))
        else:
            intervals.append(tuple(group))

    return intervals

samples_nans = samples_nans[samples_nans['id_sample'] < 5]

from datetime import datetime, timedelta

means_dict = samples_nans.groupby('link', observed=False).agg({'time_travel':[np.mean]})[('time_travel', 'mean')].to_list()

list_df_listos = [] 
for id, df_sample in samples_nans.groupby('id_sample'):
    # identificar los intervalos con nans
    df_sample.reset_index(inplace=True)
    index_nans = df_sample.index[pd.isna(df_sample['time_travel'])]
    inter_nans = intervals(list(index_nans))
    df_sample.insert(loc=2, column='id_linkref', value=df_sample.index)
    print(inter_nans)

    for gap in inter_nans:
        # generar los time travel
        means = [means_dict[i] for i in gap]

        # si hay nans en el medio, ajustamos means
        if gap[0] != 0 and gap[-1] != 82:
            time_init = df_sample.loc[gap[0] - 1, 'date_time_end']
            time_end = df_sample.loc[gap[-1] + 1, 'date_time_init']
            dif_time =  time_end - time_init
            means = list(map(lambda e: dif_time.total_seconds() * (e / sum(means)), means))

        # colocamos means
        for i in range(len(gap)):
            df_sample.loc[gap[i], 'time_travel'] = round(means[i], 0)
            if 0 not in gap:
                df_sample.loc[gap[i], 'date_time_init'] = df_sample.loc[gap[i]-1, 'date_time_end']
                df_sample.loc[gap[i], 'date_time_end'] = df_sample.loc[gap[i], 'date_time_init'] + timedelta(seconds=round(means[i], 0))
            else:
                j = (i+1)*(-1)
                df_sample.loc[gap[j], 'date_time_end'] = df_sample.loc[gap[j]+1, 'date_time_init']
                df_sample.loc[gap[j], 'date_time_init'] = df_sample.loc[gap[j], 'date_time_end'] - timedelta(seconds=round(means[i], 0))

    df_sample.drop(columns=['index', 'id_link'], inplace=True)                
    list_df_listos.append(df_sample)


# Extraer los samples con tiempo que sobrepasan el min y max permitido
min = 10
max = 1500
list_err = []
listo = []
# samplesdf = pd.concat(list_df_listos)
samplesdf = samples_nans.copy()
for id, sampledf in samplesdf.groupby('id_sample'):
    min_sample = sampledf['time_travel'].min()
    max_sample = sampledf['time_travel'].max()
    if (min_sample < min) or (max_sample > max):
        list_err.append(sampledf)
    else:
        listo.append(sampledf)

# Mostrar
print(f"{len(listo)} samples correctos")
print(f"{len(list_err)} samples con max, min trave times erroneos")

samplesdf

# verificar maximos y minimos
data = pd.concat(listo).copy().reset_index(drop=True)
print('Minimos:')
print(data.nsmallest(10, 'time_travel')['time_travel'].to_string())
print('='*20)
print('Maximos:')
print(data.nlargest(10, 'time_travel')['time_travel'].to_string())

# NANs
# verificar si hay datos faltantes
print("Cantidad de NaNs:")
for col in data:
    nans = data[col].isna().sum()
    print(f"\tColumna {col}: {nans}")

data = data[~data['time_travel'].isna()]
data

### Poner intervalos de tiempo

import pandas as pd
from collections import Counter
from datetime import datetime, time, timedelta

# Convertir las columnas de tiempo en datetime
data['date_time_init'] = pd.to_datetime(data['date_time_init'])
data['date_time_end'] = pd.to_datetime(data['date_time_end'])

# aproximar a horas las datetime (para hacer el histograma)
ix = pd.DatetimeIndex(data['date_time_end']).floor('1h')
data['time'] = ix
data['time'] = data['time'].apply(lambda dt: str(dt).split()[-1]) #

# Extraer la hora en otro columna
data['hour'] = data['date_time_end'].apply(lambda dt: dt.time().strftime("%H:%M:%S"))
data.head()

def redondear_hora(dt, split):
    m = int(round(dt.time().minute/split, 0))*split
    if m == 60 or dt.time().hour == 4: 
        return time(dt.time().hour + 1, 0)   
    else:
        return time(dt.time().hour, m)   

#  Aproximar la 'hora' en intervalos de tiempo n minutos
n = 15
name_col = 'rg' + str(n)+'min'
data[name_col] = data['date_time_end'].apply(redondear_hora, split=n)
data.head(85)

# dividir las horas y minutos en diferentes columnas
data['horas'] = data['rg15min'].apply(lambda t: t.hour)
data['minute'] = data['rg15min'].apply(lambda t: t.minute)
data

import numpy as np
means_dict = data.groupby('link', observed=False).agg({'time_travel':[np.mean]})[('time_travel', 'mean')]
std_dict = data.groupby('link', observed=False).agg({'time_travel':[np.std]})[('time_travel', 'std')]
print('promedio de travel time por link:', data['time_travel'].mean(),np.mean(means_dict.to_list()))
li = '156:157'
means_dict[li], std_dict[li]

id = data[data['horas'] == 14]['id_sample'].unique()
print(id)
tmp = np.random.choice(id)
dftmp = data[data['id_sample'] == tmp]
c = ['init_name', 'end_name', 'init_stop', 'end_stop', 'horas', 'minute', 'time_travel', 'id_sample']
print(dftmp[c].to_string())

# guardar datos listos para entrenar
data.to_csv("../data/travel_time_series-prueba.csv", index=False)
