==> code/6-recolectar_pruebas.ipynb <==
import pandas as pd

pointsdf = pd.read_csv('../data/SaveLocation_CSV_2023_07_18_09_08_47-ofi.csv', usecols=['id', 'location name', 'date', 'group name'])
pointsdf['group name'].unique()

pointsdf['datetime'] = pd.to_datetime(pointsdf['date'], format='%Y-%m-%d %H:%M:%S')
pointsdf

p1 = pointsdf[pointsdf['group name'] == 'Ida@!#ffff0000']
p1 = p1.sort_values(by='datetime')
p1.reset_index(drop=True, inplace=True)
p1

tmp1 = p1.head(p1.shape[0]-1)
tmp1.reset_index(drop=True, inplace=True)
tmp2 = p1.tail(p1.shape[0]-1)[['id','location name', 'datetime']]
tmp2.rename(columns={'id':'id 2', 'location name': 'location name 2', 'datetime': 'datetime 2'}, inplace=True)
tmp2.reset_index(drop=True, inplace=True)
tt = pd.concat([tmp1, tmp2], axis=1)

tt['time'] = tt.apply(lambda row: (row['datetime 2'] - row['datetime']).total_seconds(), axis=1)
tt.drop(columns=['group name', 'date'], inplace=True)
tt = tt[['location name', 'location name 2', 'id', 'id 2', 'datetime', 'datetime 2', 'time']]
tt['step'] = tt['id 2'] - tt['id']
tt 

from datetime import datetime, time, timedelta
def redondear_hora(dt, split):
    m = int(round(dt.time().minute/split, 0))*split
    if m == 60 or dt.time().hour == 4: 
        return time(dt.time().hour + 1, 0)   
    else:
        return time(dt.time().hour, m)

#  Aproximar la 'hora' en intervalos de tiempo n minutos
n = 15
name_col = 'rg' + str(n)+'min'
tt[name_col] = tt['datetime 2'].apply(redondear_hora, split=n)

# dividir las horas y minutos en diferentes columnas
tt['horas'] = tt['rg15min'].apply(lambda t: t.hour)
tt['minute'] = tt['rg15min'].apply(lambda t: t.minute)
tt['location name'] = tt['location name'] + ' - ' + tt['location name 2']
tt['str'] = 'min'
tt['id'] = tt['id'] - 100
tt['id 2'] = tt['id 2'] - 100
tt = tt[['location name','id','id 2','time', 'horas','minute', 'str','step']]
tt

#tt.to_csv('pruebasCampo1.csv', index=False)
